Replace webpack with gulp task
I am going to remove the 'barrel file' for the library, and replace with direct
imports, which are better for tree shaking.

As I won't be bundling JavaScript, a gulpfile will be better to manage copying
the scss.

Renaming test scss files so they all follow the same pattern, and are easier to
ignore when copying.