Add Javascript fallback styles
Reverting header styles so that the display of menu items in a vertical column (collapsible menu) is only applicable on a small screen if JavaScript is enabled.

While ordinarily I would have mobile-first styles, this was the cleanest (And easiest to understand) way of making the fallback work and be in-line with the larger screen.

Content still wraps, so it will be responsive.