Increment version number to v2.0.0
While this release will contain a relatively small change, it is a breaking one so incrementing by a major version.

Previously the use of the descriptive-list class required 3 elements per grid-row, with an icon in between the `<dt>` and <dd>`. However, this fails WCAG criteria, so removing it from the class. 

Any usages of this class would need to ensure there are only pairs of `<dt>` and `<dd>` within the `<dl>` element with the class applied.