Change module type to ESM
Without specifying type in the package.json, all files are treated as common JS as default. As use of modules in Node is stable, servers such as Vite are deprecating the use of CSM.

Changing the type here to ensure TypeScript is transpiled as ESM JavaScript.

With modules specified we need to add extensions to our imports, and remove any use of `require` import syntax.