Replace babel jest with ts-jest
Using ESM modules with .js extensions is causing tests to fail due to unresolved modules. We can handle this by transforming the files with ts-jest.

Ignore `jest.config.ts` in linting rules

Some properties have casing that does not conform, and the files uses a default export which is not recommended.