Add markdown style variant
When using the `.ld-md` class as a wrapper, underlying html elements will be applied with the same styling as those which I would normally provide class names to manage styling in html.

For the `ld-link` component, this could be applied to the parent element of an
`a` element or to the `a` itself. In the case of the parent nested scss styles
were applied to the underlying `a` element to manage hover, focus styles etc.
Need to move this handling out when the parent element could specifically be an
`a` element when introducing the markdown variant of applying styles.